fixed download

+++ ImOsmTileGrabberUrl.h
@@ -37,8 +37,8 @@ private:
   };
   std::future<GrabData> _grabFuture;
 
-  GrabData onLaunchGrab(int zmin, int zmax, int xmin, int xmax, int ymin,
-                        int ymax);
+  GrabData onLaunchGrab(int zmin, int zmax, float xmin, float xmax, float ymin,
+                        float ymax);
   Tile::Remote onHandleRequest(const std::array<int, 3> &zxy);
   static size_t onPullResponse(void *data, size_t size, size_t nmemb,
                                void *userp);
